Move `color()` function parsing for sRGB and HSV
We start with these color spaces because it is easier to move additional
parsers into sub-modules that already have support for parsing multiple
formats.
